Module starlight::database::actions::communities

Database actions concerning Communities

Functions

get_communities
Returns all of the stored communities in the Communities table

get_community_admins
Returns a vector of admins for a given Community

get_community_by_id
Returns a Community given its name (id as per supergroup spec)

put_community
Inserts a new community into the Communities table

remove_community
Removes a Community from the database

set_community_admins
Sets the admins for a new community

update_community_description
Updates the description of a given Community

update_community_title
Updates the title of a given Community