Module starlight::database::actions::post

Database actions concerning Posts

Structs

PostInformation
Struct representing an unwrapped Post, containing all useful information

Functions

get_all_posts
Returns all posts in the Posts table

get_all_top_level_posts
Returns all the top level posts stored in the Posts table

get_children_posts_of
Obtains the children of a post given a post

get_content_of_post
Obtains an array of the content of a post given a post

get_parent_of
Obtains the parent of a post given the post

get_post
Retrieves a specified post from the Posts table given its UUID

get_posts_by_user
Obtains all the posts authored by a given user

get_top_level_posts_of_community
Returns all top level posts in a specified community in the Posts table

modify_post_title
Edits the title of a Post in the Posts table

put_post
Inserts a new post and content into the database

put_post_contents
Inserts post content into the database

remove_post
Removes a post from the database

remove_post_contents
Removes the content of a post from the database

touch_post
Updates a post in the Posts table