Module starlight::database::models

Database type definitions for serialization.

Structs

DatabaseCommunitiesUser
Struct representing a row within the CommunitiesUsers table in the database

DatabaseCommunity
Struct representing a row within the Communities table in the database

DatabaseFederatedUser
Struct representing a row in the FederatedUsers table in the database

DatabaseLocalUser
Struct representing a row in the LocalUsers table in the database

DatabaseMarkdown
Struct representing a Markdown content within a Post

DatabaseNewCommunitiesUser
Struct representing a new admin to be inserted into the CommunitiesUsers table in the database

DatabaseNewCommunity
Struct representing a new Community to be added to the Communities table in the database

DatabaseNewFederatedUser
Struct representing a new FederatedUser to be inserted into the FederatedUsers table in the database

DatabaseNewLocalUser
Struct representing a new LocalUser to be inserted into the LocalUsers table in the database

DatabaseNewPost
Struct representing a new Post to be inserted into Posts table in the database

DatabaseNewUser
Struct representing a new User to be inserted into the Users table in the database

DatabasePost
Struct representing a row in the Posts table within the database

DatabaseText
Struct representing a Text content within a Post

DatabaseUser
Struct representing a row in the Users table in the database

Functions

_impl_associations_for_databasecommunitiesuser

_impl_associations_for_databasefederateduser

_impl_associations_for_databaselocaluser

_impl_associations_for_databasemarkdown

_impl_associations_for_databasepost

_impl_associations_for_databasetext

_impl_identifiable_for_databasecommunitiesuser

_impl_identifiable_for_databasecommunity

_impl_identifiable_for_databasefederateduser

_impl_identifiable_for_databaselocaluser

_impl_identifiable_for_databasemarkdown

_impl_identifiable_for_databasepost

_impl_identifiable_for_databasetext

_impl_identifiable_for_databaseuser

_impl_insertable_for_databasenewcommunitiesuser

_impl_insertable_for_databasenewcommunity

_impl_insertable_for_databasenewfederateduser

_impl_insertable_for_databasenewlocaluser

_impl_insertable_for_databasenewpost

_impl_insertable_for_databasenewuser

_impl_queryable_for_databasecommunitiesuser

_impl_queryable_for_databasecommunity

_impl_queryable_for_databasefederateduser

_impl_queryable_for_databaselocaluser

_impl_queryable_for_databasemarkdown

_impl_queryable_for_databasepost

_impl_queryable_for_databasetext

_impl_queryable_for_databaseuser