Struct starlight::database::models::DatabasePost

pub struct DatabasePost {
    pub id: u64,
    pub uuid: String,
    pub title: Option<String>,
    pub author_id: u64,
    pub created: NaiveDateTime,
    pub modified: NaiveDateTime,
    pub parent_id: Option<u64>,
    pub community_id: u64,
    pub deleted: bool,
}

Struct representing a row in the Posts table within the database

Fields
id: u64

Internal id of the Post

uuid: String

Uuid of the Post

title: Option<String>

Title of the Post (null for comments)

author_id: u64

Internal id of the author of the Post

created: NaiveDateTime

Time of Post creation

modified: NaiveDateTime

Time of latest Post modification

parent_id: Option<u64>

Internal id of the parent Post (null for top-level posts)

community_id: u64

Internal id of the community the Post belongs to

deleted: bool

Boolean representing the Post being deleted or not