Module starlight::internal::authentication

Authentication handling functions

Structs

Token
Struct representing a JWT Authentication Token

Constants

MAX_SIZE
Max request payload size of 256kBs

TIMEOUT
Timeout of one week in seconds

Statics

JWT_SECRET_KEY
JWT Secret Key used to generate Tokens.

Functions

authenticate
Authenticates a LocalUser upon login to ensure session is valid

generate_session
Returns a new UUID identifying a new session

make_federated_request
Wrapper function to send a federated request with the supergroup authentication

verify_federated_request
Wrapper method to verify an incoming request using the supergroup protocol authentication