Module starlight::internal::communities

Internal API endpoints for actions concerning communities

Structs

CommunityDetails
Struct representing a query specifying the host of a community

CommunityHost
Struct representing a community stored on a federated host

CreateCommunity
Struct representing a query to create a new community

EditCommunity
Struct representing a query to edit a community's details

ListCommunities
Struct representing a query to list the communities of an optionally specified host

SearchCommunities
Struct representing a query to search communities

create_community

delete_community

edit_community_details

get_community_details

list_communities

search_communities

Functions

external_get_community
Obtains a community stored on an external host

external_list_communities
Queries external hosts to also obtain their list of communities

local_get_community
Obtains a community that is locally stored