Module starlight::internal::posts

Internal API endpoints for actions concerning posts

Structs

CreateCommunity
Struct representing a community object sent when creating a Post

CreatePost
Struct representing the body of a POST request making a new Post

CreatePostExtern
Struct representing a body of a POST request matching that of the supergroup specification

GetPost
Struct representing a query to get all posts from a host (optionally from a community as well)

HostQuery
Struct representing the query of post creation, where a host is specified

LocatedPost
Struct representing a singular Post that can be serialized into a JSON response

SearchPosts
Struct representing a query to search posts

create_post

delete_post

edit_post

get_post

list_posts

search_posts

Functions

cache_federated_user
Writes a federated author to cache

external_get_post
Gets one post matching UUID from a host.

external_list_posts
Returns a list of all posts from one host, optionally with a specified community

external_list_posts_inner
Inner function to send a request obtaining posts from a remote host

get_post_local
Obtains a post that is locally stored

list_local_posts
Obtains all the posts stored locally, optionally filtered by a specified community

local_user_has_modify_post_permission
Determines if a user has the permissions to modify a post

request_get_post
Obtains a post that is stored on a remote host